+++ tests/linen/linen_attention_test.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Tests for flax.deprecated.nn.attention."""
+"""Tests for flax.linen.attention."""
 
 from absl.testing import absltest
 from absl.testing import parameterized